+++ NEWS.md
@@ -12,6 +12,7 @@
 - the conditional prediction interval based on `summary_heterogeneity()` is now conditional on the presence of the effect
 - additional minor prior handling fixes (i.e., missing marginal estimates when only alternative prior distributions were specified etc)
 - diagnostics with mixture baseline priors when using `algorithm = "ss"`
+- `summary_heterogeneity()` with only a single study does not produce relative heterogeneity instead of crashing
 
 ## version 3.4
 ### Features